+++ x/evm/keeper/statedb_test.go
@@ -40,7 +40,7 @@ func TestStateDBTestSuite(t *testing.T) {
 
 func (suite *StateDBTestSuite) SetupTest() {
 	suite.EVMTestSuiteWithAccountAndQueryClient.SetupTest(suite.T())
-	encodingConfig := encoding.MakeConfig()
+	encodingConfig := encoding.MakeConfig(nil)
 	suite.clientCtx = client.Context{}.WithTxConfig(encodingConfig.TxConfig)
 	suite.ethSigner = ethtypes.LatestSignerForChainID(suite.App.EvmKeeper.ChainID())
 }